+++ app/helpers/pageflow/social_share_helper.rb
@@ -51,7 +51,7 @@ module Pageflow
       ''
     end
 
-    def social_share_entry_image_tags(entry)
+    def social_share_image_urls(entry)
       image_urls = []
       image_file = ImageFile.find_by_id(entry.share_image_id)
 
@@ -68,7 +68,11 @@ module Pageflow
         end
       end
 
-      render 'pageflow/social_share/image_tags', :image_urls => image_urls
+      image_urls
+    end
+
+    def social_share_entry_image_tags(entry)
+      render 'pageflow/social_share/image_tags', image_urls: social_share_image_urls(entry)
     end
 
     def social_share_normalize_protocol(url)