+++ config/routes.rb
@@ -73,7 +73,7 @@ Pageflow::Engine.routes.draw do
   get ':entry_id/videos/:id', :to => 'files#show', :as => :short_video_file, :defaults => {:collection_name => 'video_files'}
   get ':entry_id/audio/:id', :to => 'files#show', :as => :short_audio_file, :defaults => {:collection_name => 'audio_files'}
 
-  resources :entries, :only => [:show]
+  resources :entries, only: %i(index show)
 
   get ':id', to: 'entries#show', as: :short_entry
   get ':id/embed', to: 'entries#show', defaults: {embed: '1'}, as: :entry_embed