+++ spec/features/visitor/reading_entry_spec.rb
@@ -22,4 +22,37 @@ feature 'as visitor, reading entry' do
 
     expect(Dom::Navigation.first.home_button_url).to eq('http://example.com')
   end
+
+  let(:head) { page.find 'head', visible: false }
+
+  scenario 'html head contains a link to the atom feed, when using a custom domain' do
+    theming = create(:theming, cname: 'emma.example.com')
+    account = create(:account, default_theming: theming)
+    entry = create(:entry, :published, account: account)
+
+    visit(pageflow.entry_url(entry, host: 'emma.example.com'))
+
+    atom_link_tag = head.find('link[type="application/atom+xml"]', visible: false)
+    expect(atom_link_tag['href']).to eq(pageflow.entries_url(host: 'emma.example.com', format: 'atom'))
+  end
+
+  scenario 'html head contains no link to the atom feed, when no custom domain is requested' do
+    theming = create(:theming, cname: 'emma.example.com')
+    account = create(:account, default_theming: theming)
+    entry = create(:entry, :published, account: account)
+
+    visit(pageflow.entry_url(entry))
+
+    expect(head).not_to have_selector('link[type="application/atom+xml"]', visible: false)
+  end
+
+  scenario 'html head contains no link to the atom feed, when the account does not have a custom domain' do
+    theming = create(:theming)
+    account = create(:account, default_theming: theming)
+    entry = create(:entry, :published, account: account)
+
+    visit(pageflow.entry_url(entry))
+
+    expect(head).not_to have_selector('link[type="application/atom+xml"]', visible: false)
+  end
 end