+++ spec/support/dominos/atom_feed.rb
@@ -1,0 +1,9 @@
+module Dom
+  class AtomFeed < Domino
+    selector 'feed'
+
+    def title
+      find('title')
+    end
+  end
+end