+++ CHANGELOG.md
@@ -185,3 +185,10 @@
 See
 [12-x-stable branch](https://github.com/codevise/pageflow/blob/12-x-stable/CHANGELOG.md)
 for previous changes.
+
+  ##### Notable Changes
+
+  - Atom feed
+    ([#585](https://github.com/codevise/pageflow/issues/585))
+
+  An [Atom feed](https://validator.w3.org/feed/docs/atom.html)([RFC](https://tools.ietf.org/html/rfc4287)) of published entries is now available. Every account with a `cname` will have their own feed. If a published story is protected by a password, it is excluded. Atom is a widely used standard to announce new publications. You could use it to automatically update a sidebar on a blog, for example.